refactor api client
