+++ core/network/src/commonMain/kotlin/com/stslex/core/network/api/server/client/ServerApiClient.kt
@@ -1,0 +1,5 @@
+package com.stslex.core.network.api.server.client
+
+import com.stslex.core.network.api.base.NetworkClient
+
+interface ServerApiClient : NetworkClient
