+++ core/network/src/commonMain/kotlin/com/stslex/core/network/api/server/error_handler/ErrorHandler.kt
@@ -1,0 +1,5 @@
+package com.stslex.core.network.api.server.error_handler
+
+import io.ktor.client.plugins.CallRequestExceptionHandler
+
+interface ErrorHandler : CallRequestExceptionHandler
