+++ core/network/src/commonMain/kotlin/com/stslex/core/network/api/server/http_client/ServerHttpClient.kt
@@ -1,0 +1,10 @@
+package com.stslex.core.network.api.server.http_client
+
+import io.ktor.client.HttpClient
+
+interface ServerHttpClient {
+
+    val client: HttpClient
+
+    val authClient: HttpClient
+}
