+++ core/network/src/commonMain/kotlin/com/stslex/core/network/api/server/model/ErrorRefresh.kt
@@ -1,0 +1,9 @@
+package com.stslex.core.network.api.server.model
+
+/**
+ * Error refresh token response
+ * @see ServerApiClientImpl.refreshToken
+ * @see ServerApiClientImpl.errorParser
+ * @see ServerApiClientImpl.request
+ */
+data object ErrorRefresh : Throwable()
