+++ core/network/src/commonMain/kotlin/com/stslex/core/network/api/server/model/ErrorRepeatEnd.kt
@@ -1,0 +1,8 @@
+package com.stslex.core.network.api.server.model
+
+/**
+ * Error repeat request.
+ * Show that the request was repeated after a refresh token
+ * @see ServerApiClientImpl.request
+ */
+internal data object ErrorRepeatEnd : Throwable()
